잠시파일
잠시 잠시

+++ kiosk/src/main/java/com/kiosk/headquarter/dto/store/HeadStoreProductDetailResponseDTO.java
@@ -27,7 +27,7 @@ public class HeadStoreProductDetailResponseDTO {
      * Product 엔티티의 discountRate가
      * BigDecimal이므로 동일한 타입을 사용합니다.
      */
-    private BigDecimal discountRate;
+    private Integer discountRate;
 
     private Boolean isDisplay;
